Move some functions to new lib file transaction_modules.js
